requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
scipy
mlflow
awscli
pyyaml
mlflow
lightgbm
kagglehub
boto3
dvc
nltk
imbalanced-learn
flask_cors
wordcloud
# 'dvc[s3]'